third_party/WebKit/LayoutTests/platform/win/fast/spatial-navigation/snav-clipped-overflowed-content-expected.txt




.

PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 2. Was 1.
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 3. Was 2.
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 2. Was 1.
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 1. Was start.
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test tests that a div with css overflow:auto will scroll to reveal its focusable elements
